add like tag function

+++ post/views.py
@@ -9,7 +9,9 @@ from django.views.decorators.csrf import csrf_exempt
 
 
 from rest_framework.response import Response
-from .models import Post
+from .models import Post, Like
+
+from tag.models import Tag
 
 @csrf_exempt
 @api_view(['POST'])
@@ -80,12 +82,20 @@ class PostListView(APIView):
 #### 여기까지 ###
 
     def post(self, request):
+        author = request.user
         title = request.data.get('title')
         content = request.data.get('content')
+        tag_ids = request.data.get('tags') # 태그는 리스트 형태
+        if not author.is_authenticated:
+            return Response({"detail": "Authentication credentials not provided"}, status=status.HTTP_401_UNAUTHORIZED)
         if not title or not content:
             return Response({"detail": "[title, content] fields missing."}, status=status.HTTP_400_BAD_REQUEST)
-        post = Post.objects.create(title=title, content=content)
+        for tag_id in tag_ids:
+            if not Tag.objects.filter(id=tag_id).exists():
+                return Response({"detail": "Provided tag not found."}, status=status.HTTP_404_NOT_FOUND)
+        post = Post.objects.create(title=title, content=content, author=author)
 #### 수정 ###
+        post.tags.set(tag_ids)
         serializer = PostSerializer(post)
         return Response(serializer.data, status=status.HTTP_201_CREATED)
 #### 여기까지 ###
@@ -124,7 +134,9 @@ class PostDetailView(APIView):
         try:
             post = Post.objects.get(id=post_id)
         except:
-            return Response({"detail": "Not found."}, status=status.HTTP_404_NOT_FOUND)        
+            return Response({"detail": "Not found."}, status=status.HTTP_404_NOT_FOUND)  
+        if request.user != post.author:
+            return Response({"detail": "Permission denied"}, status=status.HTTP_401_UNAUTHORIZED)      
         post.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
       
@@ -142,14 +154,43 @@ class PostDetailView(APIView):
     #         serializer.save()
     #         return Response(serializer.data, status=status.HTTP_200_OK)
 
-    def put(self, request, post_id):
+    
+
+    def patch(self, request, post_id):
         try:
             post = Post.objects.get(id=post_id)
         except:
             return Response({"detail": "Not found."}, status=status.HTTP_404_NOT_FOUND)
-        title = request.data.get('title')
-        content = request.data.get('content')
-        post = post.objects.update(title=title, content=content)
+        if request.user != post.author:
+            return Response({"detail": "Permission denied"}, status=status.HTTP_401_UNAUTHORIZED)
+        serializer = PostSerializer(post, data=request.data, partial=True)
+        if not serializer.is_valid():
+            return Response({"detail": "data validation error"}, status=status.HTTP_400_BAD_REQUEST)
+        serializer.save()
+        return Response(serializer.data, status=status.HTTP_200_OK)
+    
+class LikeView(APIView):
+    def post(self, request, post_id):
+        if not request.user.is_authenticated:
+            return Response({"detail": "Authentication credentials not provided"}, status=status.HTTP_401_UNAUTHORIZED)
+
+        ### 1 ###
+        try:
+            post = Post.objects.get(id=post_id)
+        except:
+            return Response({"detail": "Not found."}, status=status.HTTP_404_NOT_FOUND)
+
+        ### 2 ###
+        like_list = post.like_set.filter(user=request.user)
+
+        ### 3 ###
+        if like_list.count() > 0:
+            post.like_set.get(user=request.user).delete()
+        else:
+            Like.objects.create(user=request.user, post=post)
+
+        serializer = PostSerializer(instance=post)
+        return Response(serializer.data, status=status.HTTP_200_OK)
